requirements.txt
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-import-export==1.1.0
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.9
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==4.0
xlrd==1.1.0
xlwt==1.3.0
